Added some edits to song edit and view.

+++ frontend/models/Songbook_model.php
@@ -52,21 +52,22 @@ class Songbook_model extends CI_Model{
         $this->db->where('ID', $song_id);
         $this->db->update('song', $data->song_data);
         
-        if($data->song_album_data['album_id'] == 'none'){
-            $this->db->where('song_id', $song_id);
-            $this->db->delete('song_album');
-        } else {
-            $this->db->where('song_id', $song_id);
-            $this->db->delete('song_album');
-            $this->db->insert('song_album', $data->song_album_data);
+        if (isset $data->song_album_data){
+            if($data->song_album_data['album_id'] == 'none'){
+                $this->db->where('song_id', $song_id);
+                $this->db->delete('song_album');
+            } else {
+                $this->db->where('song_id', $song_id);
+                $this->db->delete('song_album');
+                $this->db->insert('song_album', $data->song_album_data);
+            }
         }
         
-        $this->db->where('song_id', $song_id);
-        $this->db->delete('lyrics');
-        $this->db->insert('lyrics', $data->lyrics_data);
-        
-        
-        
+        if (isset $data->lyrics_data){
+           $this->db->where('song_id', $song_id);
+            $this->db->delete('lyrics');
+            $this->db->insert('lyrics', $data->lyrics_data);
+        }   
     }
     
     ////////////
